Add query score info

+++ modules/rank-eval/src/main/java/org/elasticsearch/index/rankeval/HistoricalRankEvalRun.java
@@ -49,7 +49,7 @@ public record HistoricalRankEvalRun(
         return builder;
     }
 
-    record QueryResult(String queryId, float metricScore) {}
+    record QueryResult(String queryId, double metricScore) {}
 
     // This does not return metric-specific details such as relevantRatingThreshold, normalize, ignoreUnlabeled.
     // We would probably want to keep this in a non-POC.
